Added a gitolite lister which makes it possible to work with a gitolite server on a remote host
Quite a dump implementation for now, which only looks at the "expanded" output (with expand as a command to gitolite) and takes the last column as the repository name. Later one should look at the access rights too for filtering which repositories that can actually be used

gitolite also supports upload if enabled on the server, which is done simply by pushing a new project. As of now, this is only implemented with a dummy